Add options from configuration. Prevent deleting current user.

+++ Source/Fims.Server/Controllers/UserController.cs
@@ -1,3 +1,4 @@
+using System.Security.Claims;
 using Fims.Server.Data;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
@@ -13,11 +14,13 @@ public class UserController : ControllerBase
 {
   private readonly ILogger<UserController> _logger;
   private readonly UserManager<ApplicationUser> _userManager;
+  private readonly SignInManager<ApplicationUser> _signinManager;
 
-  public UserController(ILogger<UserController> logger, UserManager<ApplicationUser> userManager)
+  public UserController(ILogger<UserController> logger, UserManager<ApplicationUser> userManager, SignInManager<ApplicationUser> signinManager)
   {
     _logger = logger;
     _userManager = userManager;
+    _signinManager = signinManager;
   }
   
   [HttpGet(Name = "GetUsers")]
@@ -46,6 +49,11 @@ public class UserController : ControllerBase
   [HttpDelete]
   public async Task<IResult> DeleteUser([FromQuery] string id)
   {
+    var userId = _userManager.GetUserId(User);
+    if (userId == id)
+    {
+      return Results.BadRequest("Cannot delete current user.");
+    }
     var result = await _userManager.DeleteAsync(_userManager.Users.First(u => u.Id == id));
 
     return result.Succeeded ? Results.Ok() : Results.NotFound(result);
